backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# File processing libraries
PyPDF2==3.0.1
ebooklib==0.18.1
beautifulsoup4==4.12.2

# Text-to-speech libraries
pyttsx3==2.90
TTS==0.22.0

# Cloud storage dependencies
boto3==1.34.0
google-cloud-storage==2.10.0
google-auth==2.23.4

# Security and authentication
Flask-JWT-Extended==4.6.0
cryptography==41.0.7
bcrypt==4.1.2

# Additional utilities
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1

# Enhanced TTS with CUDA support
torch==2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0